Fix panic on select after adding a column to a table.
This is similar to #31 but it wasn't covered as it wasn't matching any case in
the switch inside file.Read as the type added to the table was a string. Doing
the same with a type like complex64, float32, int8, int16, int32, uint8,
uint16, uint32, blob, bigint, bigrat, time, duration will trigger a panic.