Fix scanning of dummy table values
This PR fixes proper scanning of rows returned  when selecting from dummy table.

Take for instance  `select 14`  .Previously we were taking the internal representation of value 14 which is idealInt(14). The driverRows.Next was not aware of the idealInt value.

This PR fix this by properly casting idealTypes (idealInt,idealFloat etc) to proper go types.