Add support for interactive mode
This commit adds `-i` flag which will drop you in an interactive shell.

You can run queries as you do with normal ql tool in the interactive shell.

At the moment there are two utility commands in the ql shell.

`\clear` which will clear your terminal( This only works on linux and darwin.

`\exit` exits the interactive sjell

`\q` is an alias for `\exit`

Eveverything else is interpreted as a ql query