Evaluate field expression when selecting on dummy tables
This commit ensures the field expression is evaluated when selecting from nothing.

For instance you can

```
select 10, now();
```

This query will return a row of two values, 10 and `time.Time`.

This functionality is handy when we try to implement the pg_sleep function to slow down queries.

so we can do something like

```
select 1;select sleep(100);select 2;
 ```