Use physical column to find simple index by name
The difference between physical and logical columns only appears after
dropping a column.

Fixing this uncovered another bug where dropped columns are returned
from selects with a WHERE clause using an index. That bug is from
dropped columns not being filtered out after reading a row. The
filtering is done in a simple table scan, that code is replicated in the
path used by index plans. This bug is also fixed.

Fixes #188